test: eliminate the 15 CI-skipped tests (pass or fail, never skip)
Codecov reported 15 tests that skip at runtime in CI. A test should pass
or fail, so each is now deterministic:

- SPA embedded-FS tests (6): the file-serving/cache-header cases used to
  probe the real embedded FS and skip when `pnpm build` had not populated
  it (the Go-only CI job never does). They duplicated the existing
  *_WithTestFS suite (backed by testStaticFS via newSPAHandler's fs.FS
  seam), so the skipping duplicates were removed.
- Live provider-API tests (3, Anthropic/OpenAI/ZAI): moved to
  discovery_live_test.go behind //go:build live. They are excluded from the
  default build entirely instead of skipping at runtime; run with
  `go test -tags live`. A missing key is now a hard failure.
- pg_dump-missing backup tests (2): removed as duplicates of
  TestCreateBackup_NoPgDump_ManipulatedPATH, which already exercises the
  412 path deterministically via a subprocess with an empty PATH.
- Proxy upstream-timeout test: un-skipped. It now shrinks request_timeout
  to 250ms via settings so a slow upstream trips it in well under a second
  (the stale skip claimed a >30s sleep; the body only sleeps 2s).
- Docker socket-not-exists test: points dockerSocketPath (already a package
  var) at a guaranteed-absent temp file instead of skipping when the host
  happens to run Docker.
- Two dead skips deleted: TestListProviders_ModelCountQueryError (a no-op
  that skipped, already covered by TestListProviders_CancelledContext) and
  TestRepository_SyncAllModels_ManuallyDisabledGroupPreserved (a skip whose
  own comment admitted the scenario did not work).

Remaining t.Skip calls are environment guards (pg_dump/pg_restore/Docker
present, non-empty catalogs) that do not fire in CI, which is why only
these 15 were reported.
