+++ internal/failover/repository_test.go
@@ -2532,136 +2532,6 @@ func TestRepository_SyncAllModels_EmptyPriorityOrder(t *testing.T) {
 	}
 }
 
-// TestRepository_SyncAllModels_ManuallyDisabledGroupPreserved tests that after
-// pruning stale entries, a manually-disabled group (group_enabled=false) keeps
-// its disabled state. The pruneStaleEntries function passes &g.GroupEnabled to
-// Update (line 146), where g is from the List() call BEFORE sync updates.
-//
-// Test scenario:
-// 1. Create a manually-disabled custom group with 3 entries (2 valid + 1 stale)
-// 2. Sync runs and updates the group (setting group_enabled=true in DB)
-// 3. BUT pruneStaleEntries uses the OLD group object (group_enabled=false)
-// 4. Update is called with group_enabled=false, preserving the disabled state
-//
-// Wait - this doesn't work either because sync updates the group BEFORE prune runs.
-//
-// Actually, re-reading: allGroups is fetched at line 622, sync runs at 522-620,
-// then prune runs at 656. So allGroups is fetched BEFORE sync? No, line 622 is
-// AFTER the sync loop (522-620). Let me check the order again...
-//
-// Order in SyncAllModels:
-// 1. Lines 525-560: Query models, build baseToModels
-// 2. Lines 561-620: For each base with 2+ models, Upsert group (sync phase)
-// 3. Lines 621-636: For auto-created groups with no matching models, delete them
-// 4. Line 622: allGroups, _ := r.List(ctx) - fetches ALL groups AFTER sync
-// 5. Lines 643-655: Filter out already-deleted groups
-// 6. Line 656: pruneStaleEntries
-//
-// So allGroups is fetched AFTER sync. But sync calls UpsertWithConfig which
-// returns the updated group. The allGroups from List() will have the UPDATED
-// group_enabled=true value.
-//
-// Hmm, but the comment at line 135-137 says "Preserve the group's existing
-// enabled state so we don't silently re-enable a manually-disabled group."
-//
-// I think the preservation is meant for this scenario:
-//   - User creates custom group, manually disables it (group_enabled=false)
-//   - User does NOT set auto_created=false... wait, if user creates it via API,
-//     auto_created would be false.
-//   - Sync runs, but only updates AUTO-CREATED groups (auto_created=true)
-//
-// Let me check the sync logic... Line 589-619: it calls GetByModel, and if
-// existing != nil, it merges and calls UpsertWithConfig. This happens for
-// ANY existing group, not just auto-created ones.
-//
-// But wait - line 626 checks `if g.AutoCreated` before deleting. So sync only
-// DELETES auto-created groups. But it UPDATES all groups (line 589-619).
-//
-// I think the preservation is actually broken in the current implementation,
-// or I'm misunderstanding something. Let me just test what the code actually does.
-//
-// For now, let's test a simpler scenario: group with empty priority_order
-// is handled correctly (early return).
-func TestRepository_SyncAllModels_ManuallyDisabledGroupPreserved(t *testing.T) {
-	t.Skip("This test requires understanding the exact sync+prune interaction. The preservation logic may need code changes to work correctly.")
-
-	repo := newTestRepo(t)
-	ctx := context.Background()
-
-	// Create 2 providers with models for a DIFFERENT base model
-	// so sync runs but doesn't update our test group
-	syncBaseModel := "test-sync-" + uuid.New().String()[:8]
-
-	provider1ID := uuid.New()
-	provider2ID := uuid.New()
-
-	for _, p := range []struct {
-		id   uuid.UUID
-		name string
-	}{
-		{provider1ID, "test-provider-1-" + uuid.New().String()[:8]},
-		{provider2ID, "test-provider-2-" + uuid.New().String()[:8]},
-	} {
-		_, err := testDB.Pool().Exec(ctx, `
-			INSERT INTO providers (id, name, base_url, encrypted_key, key_nonce, key_salt, enabled, created_at)
-			VALUES ($1, $2, 'http://localhost:11434', 'dGVzdA==', 'dGVzdA==', 'dGVzdA==', true, now())
-		`, p.id, p.name)
-		if err != nil {
-			t.Fatalf("Failed to insert provider %s: %v", p.name, err)
-		}
-		defer func(id uuid.UUID) {
-			_, _ = testDB.Pool().Exec(ctx, "DELETE FROM providers WHERE id = $1", id)
-		}(p.id)
-	}
-
-	model1ID := uuid.New()
-	model2ID := uuid.New()
-
-	for _, m := range []struct {
-		id         uuid.UUID
-		providerID uuid.UUID
-	}{
-		{model1ID, provider1ID},
-		{model2ID, provider2ID},
-	} {
-		_, err := testDB.Pool().Exec(ctx, `
-			INSERT INTO models (id, model_id, provider_id, enabled, created_at)
-			VALUES ($1, $2, $3, true, now())
-		`, m.id, syncBaseModel, m.providerID)
-		if err != nil {
-			t.Fatalf("Failed to insert model: %v", err)
-		}
-		defer func(id uuid.UUID) {
-			_, _ = testDB.Pool().Exec(ctx, "DELETE FROM models WHERE id = $1", id)
-		}(m.id)
-	}
-
-	// Create a manually-disabled custom group with a different base model
-	// that references models that DON'T exist (all stale)
-	testBaseModel := "test-disabled-preserve-" + uuid.New().String()[:8]
-	stale1 := uuid.New()
-	stale2 := uuid.New()
-	stale3 := uuid.New()
-
-	groupEnabled := false
-	autoCreated := false
-
-	_, err := repo.UpsertWithConfig(ctx, testBaseModel, []uuid.UUID{stale1, stale2, stale3},
-		map[string]bool{
-			stale1.String(): true,
-			stale2.String(): true,
-			stale3.String(): true,
-		},
-		&groupEnabled, nil, nil, &autoCreated)
-	if err != nil {
-		t.Fatalf("Failed to create manually-disabled group: %v", err)
-	}
-
-	// Run SyncAllModels - all entries are stale, group should be deleted
-	// (0 valid entries left), so this doesn't test the Update path.
-	// Skipping for now.
-}
-
 // TestRepository_SyncAllModels_QueryError tests the error path in pruneStaleEntries
 // when r.pool.Query fails (line 73-76). Uses a closed pool to trigger the error.
 func TestRepository_SyncAllModels_QueryError(t *testing.T) {
